requirements.txt
Django==3.2.6
psycopg2-binary>=2.8
celery==5.1.2
redis==3.5.3
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
django-celery-beat==2.2.1
